Move views and urls to workbasket app and update template
